twister: move test instance class out
Move test instance class into own file.
